requirements.txt
faiss-cpu
numpy
sentence-transformers
fastapi
uvicorn
pydantic
requests
python-dotenv